do_after and company will no longer break on spacemove
So I was looking into atom/movable/Move() being high up on the profile when I stumbled upon inertia_dir and had ideas.

Everything is SUPPOSE to set this to 0 when they stop space moving. I'm just gonna trust that to be the case (it is in base atom/movable/newtonian_move.

I don't want to keep calling Process_Spacemove in the loop because that feels dirty.