Makes the MC avoid triggering the byond bug that causes lag.
See http://www.byond.com/forum/?post=2021963#comment18164837

Higher fps when cpu goes over 120 causes issues. I have it scaling around 100 to give us a buffer, and not going back down until it drops below 75 to avoid it jumping too much.

This is a bit of a mess, but all of it can removed once the bug is fixed on byond's end.