Runtime's extended family doesn't spawn until the round is just about to start.
In theory this should close out the window during which Runtime would furiously breed with any other cat she can see on her tile.

In addition this lowers the chance of Not!Espeon Runtime to 5% per round. I asked Kor if I could do this and he was/is fine with it. Also added a 100 cat cap on each type of spawned cat. HOPEFULLY IT'S NEVER NEEDED.