Added preloaded drone shell dispenser to Electrical Maintenance
Drone shells will now be present roundstart, regardless if R&D are
doing their job or not.

There are drones already on the derelicht, which can easily
make their way to the mainstation anyway.

:cl: coiax
tweak: A dusty machine in maintenance has recently whirred to life, making automated drones to help fix the station.
/:cl: