Removed check preventing self target extinguisher
The comment on this code segment implies it's to stop the fire
extinguisher from going off when put in a bag. Except it doesn't go
off when put in a bag, even without this code check.

As such, it was stopping self targeting with a fire extinguisher.

Make sure the safety is off though!