Blood Type Refactor
not really any reason to rush this at the moment. But i'm touching blood
stuff anyways.

This proc is more reuseable, as it returns a list of blood types
potentially acceptable for transfusion instead of true/false, and only
takes a single argument (which means if something is checking blood type
compatibility, it only needs the type, rather than the species and blood
type of the donator etcetera, and could be used for stuff like scanners
and junk.)

It's also a bit more readable, making it easier to add new blood types
for xeno races, if ever needed.