Wisps can now be recalled with the lantern (#17999)
:cl: coiax
fix: Fixes bug where the wisp was unable to be recalled to the lantern.
/:cl:

The way that the code is structured suggests that wisp recall works in
two parts, stripping the person who's being orbited of their thermal
vision, and then putting the wisp back in. While I was there, I cleaned
up some variables.