Adds laughter demon bottle to wizard's spellbook
:cl: coiax
rscadd: Wizards have been experimenting with summoning the slightly less
lethal "laughter demon". Please be aware that these demons appear to be
as dangerous as their slaughter cousins. But adorable, nonetheless.
/:cl:

By popular demand.
Cheaper than a slaughter demon because it's not going to kill the crew,
it might even end up healing a bunch of people.