Refactors slime names; altered slime, xeno names kept when evolving
Also fixes #18352.

Slime colour and adulthood are specified in New(), which is a lot better
than having an ugly spawn(0) in the middle of a New(). Slimes also only
update their names when growing or shrinking when it matches their
existing name format.

So if you happen to be a slime called Carl Lagerfield, it won't change
your name if you grow up to adulthood.

In addition non-standard xeno names are preserved while evolving.
And it also applies your old dir to the new xeno, because that looks
slightly better.