Adds the EASYDISMEMBER and EASYLIMBATTACHMENT specflags
EASYDISMEMBER races have about double the chance to get dismembered when attacked with something that can dismember.

EASYLIMBATTACHMENT allows races with the flag to just shove limbs found lying on the floor into empty limb slots to regain their use. However this DOESN'T heal the damaged limbs, so in most cases doing this will actually add damage to the mob. Comically, the limbs attached don't need to be the originals, so these races can go around ripping off other peoples limbs and then using them for themselves.

Both these abilities are given to the the zombie and skeleton races, and constitute a significant nerf to lichs