Adds the Power Fist (#19013)
The power-fist is a semi-port of hippiestation's power-fist. Sprites are from hippiestation.

The power fist is a metallic gauntlet with a built-in piston-ram that uses an external tank as its gas source. A wrench can be used to dial up the piston's valve and make the gauntlet hit harder and farther in exchange for more gas used.

A full large-tank will get exactly 4 level-3 punches in before needing to replace the external tank. This thing is incredibly lethal for taking out singular targets at a time but very poor against crowds as odds are you're going to exhaust your supply before fending off everyone.

The TC price might be too low/high, I don't know yet until people play around with it in an actual round. All I know is punching pun-pun half-way down the escape hall in a single hit is fun.