Adds support for multiple recipe results.
Makes it so that recipes can have multiple biproducts, instead of just
one. The results list works just like the reagents list.

ex. list("reagentid" = 5, "otherreagent" = 3) where 5 and 3 are the
amounts produced.

recipe on_reaction 'created volume' is the multiplier for the recipe,
not the sum of the volume of the biproducts. Equations have been
adjusted to this change.

A couple of reactions are made slightly more efficient.

Removes a couple of trailing returns and unneccessary variable
declarations.

Scales cfl3 fireball a miniscule amount with created volume. the range
is increased by 1 for every 100 units mixed.