Adds Jamjar Glasses (#24373)
Now's PR experiments #3: Adding an item to a vending machine. Part of a series of tests for me to figure out how the code works a little better.

These are Jamjar Glasses, so called because they're as big and thick as the bottom of jam jars. They're a comical "fashion" item alternative to regular prescription glasses, with all the same functionality. They appear in the Clothesmate vending machine, where they replace one of two existing pairs of prescription glasses.
glasses

🆑
add: Nerd Co. has sent pairs of thicker prescription glasses out to Nanotrasen stations, for your local geek to wear.
/🆑