Closes up a exploit (#24543)
This was mostly harmless as it requires breaking out of control freak, but breaking out of control freak isn't hard for low security level verbs like this. I won't go into details but I was able to break out of control freak enough to run this command using a simple text editor.