Demotes HOLOGRAM to secondary flag
No difference mechanically, but frees up a bitflag slot. Reminder that
if there are no secondary flags on an object, the check is the same as
checking a var is null.

There are only a handful of hologram items in place at any time, they
don't need a full slot in flags.