Changes /obj/machinery to have atom/movable occupants
Currently `occupant` in /obj/machinery is a /mob/living type. I have
changed it to /atom/movable, and made a type cache to determine what to
"look for" when `close_machine()` is called.

This means that machines can have non-mob occupants. I have some design
ideas for this, but that's after the freeze.