Ash walkers speak draconic by default, but still know common
:cl: coiax
add: Ash walkers now know and speak Draconic by default, but still know
Galactic Common. Remember, Galcom's language key is ",0" and you can
review your known languages with the Language Menu.
/:cl:

As lizards, ash walkers should speak Draconic, but let's not deprive
them of the chance to communicate with the smelly humans. It will just
be up to the ash walkers whether they want actually talk to the humans
or not.

Also acts as a teaching experience for people about how to use their
language menus.