Annoyed admirals recalling shuttles
:cl: coiax
add: When a shuttle is called, sometimes an on-call admiral, using
availible information to them, will recall the shuttle from Centcom.
/:cl:

This is in essence, a simple IC way for admins to recall the shuttle
if they feel that the shuttle should not come. It's up to them entirely,
the system does not trigger otherwise. They also get to set a cranky
message that is automatically fed back to the crew.

It would be nice if admins used their own judgement if the crew is
evacuating for no real reason, and might send an ERT at the same time to
keep the round going? Who knows, I'm just providing the tools.