isatom is now an alias for isloc (#32514)
isloc() returns true on anything that can be a loc. Only atoms can be a loc. Its basically isatom, but with a odd ass name.

It, like other builtin byond is* procs, uses typeid to check, not istype, saving the standard mess of recursive type searching that istype uses, thus its faster.