Fake nuclear disks are even more convincing (#34466)
* Fake nuclear disks are even more convincing

:cl: coiax
add: Fake nuclear disks can only be identified by the captain,
observers, nuclear operatives, seeing where the pinpointer points, or
attempting to put it into a nuclear device.
add: Fake nuclear disks "respawn" on the station just like the real one.
/:cl:

- Removes anti-duplication on the syndicate nuke, gets in the way of
testing.

Why? If you manage to pickpocket the Head of Security, steal the disk,
and replace it with a fake, then he shouldn't be able to tell the
difference. Is the disk you're holding real? Maybe you should pop over
to the Vault to check. As a nuke op, leave a fake disk near the very
real, and very armed nuke, as a final taunt to the crew.