Blood contract knives only last two minutes, like the chainsaw (#42074)
* Added admin smite: Mark for Death

:cl: coiax
admin: Admins can Mark For Death with the Smite menu, which functions
the same as someone being blood contracted.
tweak: The butcher knives and the red colouring of someone
marked for death will disappear after two minutes.
/:cl:

The primary use I see of Mark for Death? WGW readers. Saw some
other admins use it, but they had to manually spawn in a contract
and use it.

The cleaning up after two minutes is so the station isn't covered in
cleavers, and the RED GET THIS GUY fades, so people can be horrified/thrilled
at their actions appropriately.

* No more smite for death

* Also remove that line

* Only the knives