MakeDirty() is only called 33% of the time (#42233)
I randomly thought of this. On Metastation there are ~44,000 floor
turfs being initialized, and by moving the prob for dirtiness out of the
proc, we should be able to avoid about 2/3 of those (so around ~29,000).

I don't think I've ever done profiling, so I'm unclear how to test it.

But it seemed like maybe a good idea?