Improves Spiritual quirk
:cl: coiax
add: Players with the Spiritual quirk now spawn with a pack of candles
and a box of matches, for better communion with their chosen power, and
get a mood boost while near a holy person (generally the chaplain).
/:cl:

Makes Spiritual not just a quirk that literally requires admin
approval to pay off. Have some free candles, talk to the chaplain to
cheer up, and maybe you'll be able to gang together to convince the gods
to give you an angry bloodcrawling duck.