Mime Foods/Drinks Now Silence the Mime (#42499)
Mime stuff silences the mime now
cl
tweak: The nothing, silencer, and blank paper beverages now apply a mute on the mime for their duration (and a bit after).
cl

If you want to heal as a mime, you should not be allowed to speak. By extension things such as mime burgers will have this effect given the fact that they have nothing inside them (the reagent).