Add KDevelop and Kate to gitignore (#43367)
Added the following to .gitignore

*.kdev4*
*.kate-swp

I've already got both of these in my ~/.config/git/ignore file, but this
may be helpful to others. Since I've already got it in my global gitignore
file, I am not affected whether this is merged or not. This change was made
in the interest in other people who use or want to use KDevelop and have
not set this up in their global git ignore file.