Move the player's selection items to the top (#44868)
The chef's compartment and any future compartments of this type are
moved to the top of the interface.

It is time to unleash the awesome power of the free market and
cultivate a spirit of entrepreneurship. By placing the chef's
compartment on top of the subsidized swill provided by tyrannical
pseudo-governmental entity know as Nanotrasen, we can promote the
commerce of those possessing a mind undulled by the poison known as
socialism. God Bless the USA