fixes all the bugs you probably didnt know about (#46264)
About The Pull Request

As mentioned in codebus with the recent patch for circuits being able to produce any item (see BeeStation/BeeStation-Hornet#345), people often make a mistake in attempting to check if a collection does not contain an element. The proper execution of such a check, following the attempted formatting, would be...

!(x in y)

But instead we have lots of

!x in y

In other words, 1 or 0 in collection y, not good!
Why It's Good For The Game

Fixes a lot of bugs that likely nobody has ever noticed, probably introduces features that were intended but incorrectly coded. I have attempted to summarize what are probably the effects of this change below. I've moved interesting fixes to the top of this list.

    Dynamic mode ruleset should no longer ignore player preferences when selecting antagonist candidates.
    Pet carriers should now properly cancel callbacks for a mob escaping the carrier if they are no longer an occupant of it.
    Eightballs should now prevent ghosts from voting on answers that are not expected by the eightball.
    Modifying variables in view variables should now prevent you from adding a non-existent variable to a datum.
    The Herald's Beacon should no longer attempt to remove a non-existent voter from its list of users who need to vote. (Likely prevents a runtime)

Changelog

cl bobbahbrown
fix: Dynamic mode ruleset will now respect your player preferences when selecting antag candidates
code: Fixed 9 instances of incorrect not-in-list expressions.
/cl